requirements.txt
asgiref==3.8.1
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.32.0
azure-mgmt-core==1.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.12
django-cors-headers==3.7.0
django-filter==24.3
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
Markdown==3.7
msrest==0.7.1
mssql-django==1.5
oauthlib==3.2.2
packaging==24.2
Pillow==9.1.1
psycopg2==2.9.1
PyJWT==2.1.0
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2021.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==5.3.0
